nyit/RApro/sunlink/terraform-cilium-python/requirements.txt
click==8.1.3
python-dotenv==1.0.0
